requirements.txt
ijson==3.4.0
matplotlib==3.10.5
numpy==2.2.6
pandas==2.3.2
pymongo==4.15.1
python-dotenv==1.0.1
scikit-learn==1.7.1
scipy==1.15.3
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.0
fonttools==4.59.1
joblib==1.5.1
kiwisolver==1.4.9
packaging==25.0
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0